+++ master/libspider/papify/PapifyAction.h
@@ -61,7 +61,7 @@ public:
      * @param monitorTime       Precise if papify should also monitor time for this actor
      */
     PapifyAction(const char *PEType,
-                 long long PEId,
+                 const char *PEId,
                  const char *actorName,
                  int numberOfEvents,
                  std::vector<const char *> &moniteredEventSet,
@@ -75,7 +75,7 @@ public:
      * @param papifyAction  Original PapifyAction
      * @param PEId          The PE id
      */
-    PapifyAction(PapifyAction &papifyAction, long long PEId);
+    PapifyAction(PapifyAction &papifyAction, const char *PEId);
 
     ~PapifyAction();
 
@@ -113,7 +113,7 @@ public:
     }
 
 private:
-    long long PEId_;
+    const char *PEId_;
     const char *PEType_;
     const char *actorName_;